Project/backend/requirements-clean.txt
# Django Backend Requirements
# Gerado automaticamente - use 'pip install -r requirements.txt' para instalar

# Framework principal
Django==5.2.6

# API REST Framework
djangorestframework==3.16.1

# CORS para permitir requisições do frontend
django-cors-headers==4.9.0

# Dependências do Django (instaladas automaticamente)
asgiref==3.9.2
sqlparse==0.5.3
tzdata==2025.2